Fixed a number of bugs in the NCR 53C810 code; 64-bit Windows 86Box now compiles without warnings.

+++ src/network/slirp/cksum.c
@@ -72,7 +72,7 @@ int cksum(struct SLIRPmbuf *m, int len)
 	/*
 	 * Force to even boundary.
 	 */
-	if ((1 & (long) w) && (mlen > 0)) {
+	if ((1 & (intptr_t) w) && (mlen > 0)) {
 		REDUCE;
 		sum <<= 8;
 		s_util.c[0] = *(u_int8_t *)w;
